move types to int tests
