+++ src/constants.ts
@@ -2,20 +2,3 @@ export const NEUTRON_DENOM = process.env.NEUTRON_DENOM || 'untrn';
 export const ADMIN_MODULE_ADDRESS =
   'neutron1hxskfdxpp5hqgtjj6am6nkjefhfzj359x0ar3z';
 
-/**
- * @deprecated since version 0.5.0
- */
-export const IBC_ATOM_DENOM = process.env.IBC_ATOM_DENOM || 'uibcatom';
-/**
- * @deprecated since version 0.5.0
- */
-export const IBC_USDC_DENOM = process.env.IBC_USDC_DENOM || 'uibcusdc';
-/**
- * @deprecated since version 0.5.0
- */
-export const COSMOS_DENOM = process.env.COSMOS_DENOM || 'uatom';
-/**
- * @deprecated since version 0.5.0
- */
-export const IBC_RELAYER_NEUTRON_ADDRESS =
-  'neutron1mjk79fjjgpplak5wq838w0yd982gzkyf8fxu8u';
