+++ src/types.ts
@@ -380,60 +380,3 @@ export class Wallet {
 
 export type CodeId = number;
 
-// DenomTraceResponse is the response model for the ibc transfer denom trace query.
-// TODO: use rpc authogeneration
-/**
- * @deprecated since version 0.5.0
- */
-export type DenomTraceResponse = {
-  path?: string;
-  base_denom?: string;
-};
-
-// TODO: use rpc authogeneration
-/**
- * @deprecated since version 0.5.0
- */
-export type TotalSupplyByDenomResponse = {
-  amount: Coin;
-};
-
-// TODO: use rpc authogeneration
-/**
- * @deprecated since version 0.5.0
- */
-export type DenomMetadataResponse = {
-  metadatas: [
-    {
-      description: string;
-      denom_units: [
-        {
-          denom: string;
-          exponent: number;
-          aliases: [string];
-        },
-      ];
-      base: string;
-      display: string;
-      name: string;
-      symbol: string;
-      uri: string;
-      uri_hash: string;
-    },
-  ];
-  pagination: {
-    next_key: string;
-    total: string;
-  };
-};
-
-// TODO: use rpc authogeneration
-/**
- * @deprecated since version 0.5.0
- */
-// TotalBurnedNeutronsAmountResponse is the response model for the feeburner's total-burned-neutrons.
-export type TotalBurnedNeutronsAmountResponse = {
-  total_burned_neutrons_amount: {
-    coin: Coin;
-  };
-};
